refactor(api): extract settings routes to src/api/settings_routes.py
Move 4 settings routes (LLM GET/PUT, data-sources GET/PUT) and
7 exclusive helpers from api_server.py into a self-contained
route module following the PR #375 pattern.

- settings_routes.py defines all Pydantic models locally (no shared modules)
- register_settings_routes() resolves auth deps via sys.modules fallback
- Late-access closures for monkeypatch targets (ENV_PATH, ENV_EXAMPLE_PATH)
- Re-exports: ENV_PATH, ENV_EXAMPLE_PATH, _baostock_supported,
  _baostock_installed, _load_llm_providers for test compatibility
- api_server.py: -294 lines (3449 -> 3155)

AI-Model: idealab/Qwen3.7-Max-DogFooding
AI-Contributed/Feature: 0/756
AI-Contributed/UT: 0/0
