+++ agent/api_server.py
@@ -138,72 +138,8 @@ class RunResponse(BaseModel):
     run_logs: Optional[List[Dict[str, Any]]] = Field(None, description="Structured stdout/stderr lines")
 
 
-class LLMProviderOption(BaseModel):
-    """Supported LLM provider metadata for the settings UI."""
-
-    name: str
-    label: str
-    api_key_env: Optional[str] = None
-    base_url_env: str
-    default_model: str
-    default_base_url: str
-    api_key_required: bool = True
-    auth_type: str = "api_key"
-    login_command: Optional[str] = None
 
 
-class LLMSettingsResponse(BaseModel):
-    """Current LLM runtime settings."""
-
-    provider: str
-    model_name: str
-    base_url: str
-    api_key_env: Optional[str] = None
-    api_key_configured: bool
-    api_key_hint: Optional[str] = None
-    api_key_required: bool
-    temperature: float
-    timeout_seconds: int
-    max_retries: int
-    reasoning_effort: str
-    sse_timeout_seconds: int
-    env_path: str
-    providers: List[LLMProviderOption]
-
-
-class UpdateLLMSettingsRequest(BaseModel):
-    """Update LLM settings persisted to agent/.env."""
-
-    provider: str = Field(..., min_length=1)
-    model_name: str = Field(..., min_length=1)
-    base_url: Optional[str] = None
-    api_key: Optional[str] = None
-    clear_api_key: bool = False
-    temperature: float = 0.0
-    timeout_seconds: int = Field(120, ge=1, le=3600)
-    max_retries: int = Field(2, ge=0, le=20)
-    reasoning_effort: Optional[str] = None
-
-
-class DataSourceSettingsResponse(BaseModel):
-    """Current data source credential settings."""
-
-    tushare_token_configured: bool
-    tushare_token_hint: Optional[str] = None
-    baostock_supported: bool
-    baostock_installed: bool
-    baostock_message: str
-    env_path: str
-
-
-class UpdateDataSourceSettingsRequest(BaseModel):
-    """Update project-local data source credentials."""
-
-    tushare_token: Optional[str] = None
-    clear_tushare_token: bool = False
-
-
-# ---- V4 Session Models ----
 
 class CreateSessionRequest(BaseModel):
     """Create session request body."""
@@ -965,32 +901,6 @@ async def require_settings_write_auth(
 # Helper Functions
 # ============================================================================
 
-LLM_PROVIDER_CONFIG_PATH = AGENT_DIR / "src" / "providers" / "llm_providers.json"
-
-
-def _load_llm_providers() -> List[LLMProviderOption]:
-    """Load provider metadata from JSON so additions stay data-driven."""
-    try:
-        raw = json.loads(LLM_PROVIDER_CONFIG_PATH.read_text(encoding="utf-8"))
-        providers = [LLMProviderOption(**item) for item in raw]
-    except Exception as exc:
-        raise RuntimeError(f"Failed to load LLM provider config: {LLM_PROVIDER_CONFIG_PATH}") from exc
-
-    seen: set[str] = set()
-    for provider in providers:
-        if provider.name in seen:
-            raise RuntimeError(f"Duplicate LLM provider name: {provider.name}")
-        seen.add(provider.name)
-    if not providers:
-        raise RuntimeError("LLM provider config must not be empty")
-    return providers
-
-
-LLM_PROVIDERS = _load_llm_providers()
-LLM_PROVIDER_BY_NAME = {provider.name: provider for provider in LLM_PROVIDERS}
-LLM_REASONING_EFFORTS = {"", "low", "medium", "high", "max"}
-LLM_API_KEY_PLACEHOLDERS = {"", "sk-or-v1-your-key-here", "sk-xxx", "xxx", "gsk_xxx"}
-TUSHARE_TOKEN_PLACEHOLDERS = {"", "your-tushare-token"}
 
 
 def _ensure_agent_env_file() -> Path:
@@ -1024,19 +934,6 @@ def _read_env_values(path: Path) -> Dict[str, str]:
         if key:
             values[key] = _strip_env_value(value)
     return values
-
-
-def _read_settings_env_values() -> Dict[str, str]:
-    """Read settings without creating agent/.env.
-
-    Prefer the user's active agent/.env. If it does not exist yet, fall back to
-    agent/.env.example for display defaults only.
-    """
-    if ENV_PATH.exists():
-        return _read_env_values(ENV_PATH)
-    if ENV_EXAMPLE_PATH.exists():
-        return _read_env_values(ENV_EXAMPLE_PATH)
-    return {}
 
 
 def _project_relative_path(path: Path) -> str:
@@ -1104,105 +1001,6 @@ def _coerce_int(value: str, default: int) -> int:
         return int(value)
     except (TypeError, ValueError):
         return default
-
-
-def _build_llm_settings_response(values: Optional[Dict[str, str]] = None) -> LLMSettingsResponse:
-    """Build the public settings payload from dotenv values."""
-    env_values = values if values is not None else _read_settings_env_values()
-    provider_name = env_values.get("LANGCHAIN_PROVIDER", "openai").strip().lower()
-    provider = LLM_PROVIDER_BY_NAME.get(provider_name, LLM_PROVIDER_BY_NAME["openai"])
-    api_key = env_values.get(provider.api_key_env or "", "") if provider.api_key_env else ""
-    api_key_configured = _is_configured_secret(api_key, LLM_API_KEY_PLACEHOLDERS)
-    api_key_hint = None
-    if provider.auth_type == "oauth":
-        try:
-            from src.providers.openai_codex import get_openai_codex_login_status
-
-            token = get_openai_codex_login_status()
-        except Exception:
-            token = None
-        api_key_configured = bool(token)
-        api_key_hint = None
-    return LLMSettingsResponse(
-        provider=provider.name,
-        model_name=env_values.get("LANGCHAIN_MODEL_NAME", provider.default_model),
-        base_url=env_values.get(provider.base_url_env, provider.default_base_url),
-        api_key_env=provider.api_key_env,
-        api_key_configured=api_key_configured,
-        api_key_hint=api_key_hint,
-        api_key_required=provider.api_key_required,
-        temperature=_coerce_float(env_values.get("LANGCHAIN_TEMPERATURE", "0.0"), 0.0),
-        timeout_seconds=_coerce_int(env_values.get("TIMEOUT_SECONDS", "120"), 120),
-        max_retries=_coerce_int(env_values.get("MAX_RETRIES", "2"), 2),
-        reasoning_effort=env_values.get("LANGCHAIN_REASONING_EFFORT", "").strip().lower(),
-        sse_timeout_seconds=_coerce_int(env_values.get("VIBE_TRADING_SSE_TIMEOUT", "90"), 90),
-        env_path=_project_relative_path(ENV_PATH),
-        providers=LLM_PROVIDERS,
-    )
-
-
-def _baostock_supported() -> bool:
-    """Check whether the project has a BaoStock loader implementation."""
-    loader_dir = AGENT_DIR / "backtest" / "loaders"
-    return any((loader_dir / name).exists() for name in ("baostock.py", "baostock_loader.py"))
-
-
-def _baostock_installed() -> bool:
-    """Check whether the optional BaoStock package is importable."""
-    import importlib.util
-
-    return importlib.util.find_spec("baostock") is not None
-
-
-def _build_data_source_settings_response(values: Optional[Dict[str, str]] = None) -> DataSourceSettingsResponse:
-    """Build the public data source settings payload."""
-    env_values = values if values is not None else _read_settings_env_values()
-    token = env_values.get("TUSHARE_TOKEN", "")
-    token_configured = _is_configured_secret(token, TUSHARE_TOKEN_PLACEHOLDERS)
-    supported = _baostock_supported()
-    installed = _baostock_installed()
-    if supported:
-        baostock_message = "BaoStock loader is available."
-    elif installed:
-        baostock_message = "BaoStock package is installed, but this project has no BaoStock loader."
-    else:
-        baostock_message = "No BaoStock loader is registered in this project."
-    return DataSourceSettingsResponse(
-        tushare_token_configured=token_configured,
-        tushare_token_hint=None,
-        baostock_supported=supported,
-        baostock_installed=installed,
-        baostock_message=baostock_message,
-        env_path=_project_relative_path(ENV_PATH),
-    )
-
-
-def _sync_runtime_env(provider: LLMProviderOption, updates: Dict[str, str]) -> None:
-    """Apply saved LLM settings to the running API process."""
-    for key, value in updates.items():
-        if value:
-            os.environ[key] = value
-        else:
-            os.environ.pop(key, None)
-
-    if provider.api_key_env:
-        key_value = os.environ.get(provider.api_key_env, "")
-        if _is_configured_secret(key_value, LLM_API_KEY_PLACEHOLDERS):
-            os.environ["OPENAI_API_KEY"] = key_value
-        else:
-            os.environ.pop("OPENAI_API_KEY", None)
-    elif provider.auth_type == "oauth":
-        os.environ.pop("OPENAI_API_KEY", None)
-    else:
-        os.environ["OPENAI_API_KEY"] = "ollama"
-
-    base_url = os.environ.get(provider.base_url_env, "")
-    if base_url:
-        os.environ["OPENAI_API_BASE"] = base_url
-        os.environ["OPENAI_BASE_URL"] = base_url
-    else:
-        os.environ.pop("OPENAI_API_BASE", None)
-        os.environ.pop("OPENAI_BASE_URL", None)
 
 
 def _load_json_file(path: Path) -> Optional[Dict[str, Any]]:
@@ -1600,112 +1398,6 @@ async def list_runs(limit: int = 20):
         ))
 
     return results
-
-
-@app.get(
-    "/settings/llm",
-    response_model=LLMSettingsResponse,
-    dependencies=[Depends(require_local_or_auth)],
-)
-async def get_llm_settings():
-    """Return project-local LLM settings for the Web UI."""
-    return _build_llm_settings_response()
-
-
-@app.put("/settings/llm", response_model=LLMSettingsResponse, dependencies=[Depends(require_settings_write_auth)])
-async def update_llm_settings(payload: UpdateLLMSettingsRequest):
-    """Persist project-local LLM settings and update the running process."""
-    provider_name = payload.provider.strip().lower()
-    provider = LLM_PROVIDER_BY_NAME.get(provider_name)
-    if provider is None:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Unsupported LLM provider")
-
-    model_name = payload.model_name.strip()
-    if not model_name:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Model name is required")
-
-    if payload.temperature < 0 or payload.temperature > 2:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Temperature must be between 0 and 2")
-
-    reasoning_effort = (payload.reasoning_effort or "").strip().lower()
-    if reasoning_effort not in LLM_REASONING_EFFORTS:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Reasoning effort must be low, medium, high, or max")
-
-    current_values = _read_settings_env_values()
-    base_url = (payload.base_url if payload.base_url is not None else provider.default_base_url).strip()
-    if provider.auth_type == "oauth":
-        try:
-            from src.providers.openai_codex import validate_codex_base_url
-
-            base_url = validate_codex_base_url(base_url)
-        except ValueError as exc:
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
-    updates: Dict[str, str] = {
-        "LANGCHAIN_PROVIDER": provider.name,
-        "LANGCHAIN_MODEL_NAME": model_name,
-        provider.base_url_env: base_url,
-        "LANGCHAIN_TEMPERATURE": str(payload.temperature),
-        "TIMEOUT_SECONDS": str(payload.timeout_seconds),
-        "MAX_RETRIES": str(payload.max_retries),
-    }
-    if reasoning_effort or "LANGCHAIN_REASONING_EFFORT" in current_values:
-        updates["LANGCHAIN_REASONING_EFFORT"] = reasoning_effort
-
-    if provider.api_key_env:
-        if payload.clear_api_key:
-            updates[provider.api_key_env] = ""
-        elif payload.api_key is not None and payload.api_key.strip():
-            api_key = payload.api_key.strip()
-            updates[provider.api_key_env] = api_key if _is_configured_secret(api_key, LLM_API_KEY_PLACEHOLDERS) else ""
-        elif provider.api_key_env in current_values and _is_configured_secret(
-            current_values[provider.api_key_env],
-            LLM_API_KEY_PLACEHOLDERS,
-        ):
-            updates[provider.api_key_env] = current_values[provider.api_key_env]
-    elif payload.clear_api_key:
-        os.environ.pop("OPENAI_API_KEY", None)
-
-    _write_env_values(ENV_PATH, updates)
-    _sync_runtime_env(provider, updates)
-    return _build_llm_settings_response(_read_env_values(ENV_PATH))
-
-
-@app.get(
-    "/settings/data-sources",
-    response_model=DataSourceSettingsResponse,
-    dependencies=[Depends(require_local_or_auth)],
-)
-async def get_data_source_settings():
-    """Return project-local data source credentials for the Web UI."""
-    return _build_data_source_settings_response()
-
-
-@app.put(
-    "/settings/data-sources",
-    response_model=DataSourceSettingsResponse,
-    dependencies=[Depends(require_settings_write_auth)],
-)
-async def update_data_source_settings(payload: UpdateDataSourceSettingsRequest):
-    """Persist project-local data source credentials and update the running process."""
-    current_values = _read_settings_env_values()
-    updates: Dict[str, str] = {}
-
-    if payload.clear_tushare_token:
-        updates["TUSHARE_TOKEN"] = ""
-    elif payload.tushare_token is not None and payload.tushare_token.strip():
-        updates["TUSHARE_TOKEN"] = payload.tushare_token.strip()
-    elif "TUSHARE_TOKEN" in current_values:
-        updates["TUSHARE_TOKEN"] = current_values["TUSHARE_TOKEN"]
-
-    if updates:
-        _write_env_values(ENV_PATH, updates)
-        token = updates.get("TUSHARE_TOKEN", "").strip()
-        if _is_configured_secret(token, TUSHARE_TOKEN_PLACEHOLDERS):
-            os.environ["TUSHARE_TOKEN"] = token
-        else:
-            os.environ.pop("TUSHARE_TOKEN", None)
-
-    return _build_data_source_settings_response(_read_env_values(ENV_PATH))
 
 
 # ============================================================================
@@ -2203,6 +1895,21 @@ register_system_routes(app)
 
 # Re-export for test monkeypatch compatibility
 from src.api.system_routes import _terminate_current_process  # noqa: F401, E402
+
+
+# ============================================================================
+# Settings routes - defined in src/api/settings_routes.py
+# ============================================================================
+
+from src.api.settings_routes import register_settings_routes  # noqa: E402
+register_settings_routes(app)
+
+# Re-export for test monkeypatch compatibility
+from src.api.settings_routes import (  # noqa: F401, E402
+    _baostock_supported,
+    _baostock_installed,
+    _load_llm_providers,
+)
 
 
 # ============================================================================
